* Resolved back button issue * Updated home screen * Code enhancement
