+++ src/components/Tree.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {Button, Subheading, withTheme} from 'react-native-paper';
+import {Subheading, withTheme} from 'react-native-paper';
 import {StyleSheet, View} from 'react-native';
 import FastImage from 'react-native-fast-image';
 
@@ -9,21 +9,12 @@ const Tree = props => {
     const { colors } = props.theme;
     return (
       <View
-        style={{
-          flex: 1,
-          justifyContent: 'center',
-          alignItems: 'center',
-        }}>
+        style={styles.container}>
         <FastImage
           source={require('../assets/tree.png')}
           style={[styles.icons, {backgroundColor: colors.background}]}
         />
         <Subheading>{props.message}</Subheading>
-        <View>
-          <Button icon="refresh" mode="outlined" onPress={props.fetchData}>
-            Refresh
-          </Button>
-        </View>
       </View>
     );
 };
@@ -39,6 +30,10 @@ const styles = StyleSheet.create({
     width: 209,
     height: 220,
     borderRadius: 4,
+  },
+  container: {
+    justifyContent: 'center',
+    alignItems: 'center',
   },
 });
